Refactoring including switching FixedTimeZone to a concrete type.
Previously we had OffsetTimeZone and DaylightSavingTimeZone as
subtypes of FixedTimeZone to avoid having to store the DST offset
for fixed timezones that didn't include this information. I don't
believe that the minor memory savings is worth the trouble currently
so I've remove the original two concrete types in favor of a single
type FixedTimeZone.

Additional changes:
- "dst_offset" and "utc_offset" removed.
- "total_offset" renamed to "offset".
- Added "localtime" and "utc" accessors for extracting a DateTime
- Removed sizehint! that wasn't necessary.