Bugfix for fixed timezones using tzfile version 2
Version 2 files include a single transition time which was causing
read_tzfile to treat the FixedTimeZone as a VariableTimeZone. I've put
an additional testcase in place which tests for this particular case.

Additionally, reading a FixedTimeZone or VariableTimeZone now uses the
name passed in rather than using the timezone abbreviation in the case
of the FixedTimeZone.

Note that this bug was discovered by the localzone test using the system
installed file "Etc/GMT-9".