Simplify order_rules by putting cutoff in resolve!
We can simplify determining the cutoff by just running order_rules for
one additional year after the max_year. Changing the code in the way
also allows us to switch the cutoff to be a datetime which means that
you can generate valid ZonedDateTime in a TimeZone up to when the next
transition should start.