Fix issue with parsing abbreviation pair
When parsing the FORMAT column from the tz database there is sometimes
a pair of strings separated by a slash (‘/’). When this occurs the first
string is the abbreviation for standard time and the second string is
the abbreviation for daylight saving time.

Reference: ftp://ftp.iana.org/tz/tz-how-to.html