Improved performance of repeated use of TimeZone
Previously when using `TimeZone(x)` a serialized `TimeZone` was read
from disk and deserialized. When performing tasks such as parsing the
same `TimeZone` may be requested multiple times which would be slowed
down by the deserialization process. Now once a `TimeZone` has been
loaded it is stored in a Dict such that repeated instantiations are
fast.

Additionally this has the added benefit that
`TimeZone(x) === TimeZone(x)` since `FixedTimeZone` and
`VariableTimeZone` are immutable.