Add documentation related to canonicalization
Note that I'm not entirely consistent with how I expect things to work
when it comes to arithmetic with multiple periods. This is somewhat a
mixture of my personal taste and how Date and period arithmetic works
in base Julia. Base Julia states:

> ... in the presence of multiple Periods, the operations will be
> ordered by the Periods' types, not their value or positional order;
> this means Year will always be added first, then Month, then Week, etc

The logic works well for `Date` and `DateTime` but does seem to slightly
break down with `ZonedDateTime` due to the differences between
`DatePeriod` and `TimePeriod` math.