Support first TZ format
When extracting time zone information from the TZ environmental variable
we previously only supported loading information from a file. We now
also support the `std offset` specification which always produces a
FixedTimeZone.

The introduction of this change was required by a change to the
Travis CI environment where the time zone is now specified as `TZ="UTC"`