Revise support for ON rules (Fri<=1)
The 2019b tzdata used an ON rule "Fri<=1" which revealed that our
handling of `<=` rules to be incorrect. Our handling of `>=` rules
appears to always have been handled correctly.

Previously the algorithm assumed that the resulting date would always be
within the IN month which turned out to be invalid and major
modifications to internal functions were required to properly support
this.