Use `Scratch` for cached timezone data.
Fixes #338. At module init runs the current `build` to generate
the cached data, and stores it in a scratchspace instead of the
deps folder. Associated globals that were defined at top-level are
now set in `__init__` (and `_init` for submodules) to allow for
`PackageCompiler` support. Scratchspace is linked to the Julia and
package version so should rebuild when those change for users.

The `deps/build.jl` now just imports the package which triggers
a build of the cache if needed so this shouldn't really impact
`using TimeZones` much as far as I could tell.

A required PR for `Scratch` to expand it's version support is at
https://github.com/JuliaPackaging/Scratch.jl/pull/24. I've enabled
support going back to 1.0 for that package to match what `TimeZones`
says it supports, but it turns out `EzXML` is bounded at `1.3` so
I couldn't test down to 1.0-1.2.