Use `Scratch` to hold downloads and compiled tzdata
We don't want to mutate the package directory; let's use a scratchspace
instead, and adjust all `const` variables to do runtime lookups, for
maximum relocatability.

We shift around the data a bit here, to make it more clear that the
package directory is not to be mutated.  The ideal case here now is
that, on a fresh install, nothing is downloaded for the default happy
path, the package directory is never mutated, and the build step is
minimal in cost.